girnode: Improve int types in GIIrNodeUnion
`GIIrNodeUnion` is built dynamically at runtime (rather than being
mmapped from disk), so its types can accurately reflect their runtime
semantics, rather than an on-disk format.

As part of this, switch from `atoi()` to `g_ascii_string_to_unsigned()`
for parsing the relevant fields from a GIR XML file. This means we now
get error handling for invalid integers.


Helps: #3155

+++ girepository/girnode-private.h
@@ -384,7 +384,7 @@ struct _GIIrNodeUnion
   size_t size;
   GIIrOffsetsState offsets_state;
 
-  int discriminator_offset;
+  size_t discriminator_offset;
   GIIrNodeType *discriminator_type;  /* (owned) */
 };
